requirements.txt
<<<<<<< HEAD
django==5.2.4
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
asgiref==3.9.1
sqlparse==0.5.3
tzdata==2025.2
djangorestframework==3.16.0
=======
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
pydantic==2.11.7
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
alembic==1.16.4
pydantic-settings==2.10.1
fastapi[all]

>>>>>>> b3bb375a02daeb3b0f471dd3074c748e9c692b2d
